Reorganize components + fix naming
