+++ src/sections/hero.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import React from "react";
 
-import Navbar from "./navbar";
+import Navbar from "../components/hero/navbar";
 
 export default function Hero() {
   return (
